@@ -18,11 +18,11 @@
         }
     ],
     "require": {
-        "php": ">=7.1",
+        "php": ">=7.4 <8.0",
         "squizlabs/php_codesniffer": "^3.6"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.5"
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-4": {
@@ -48,6 +48,6 @@
     "scripts": {
         "test": "vendor/bin/phpunit",
         "phpcs": "vendor/bin/phpcs -p",
-        "coverage": "debug vendor/bin/phpunit --coverage-html=build/coverage"
+        "coverage": "xdebug-coverage vendor/bin/phpunit --coverage-html=build/coverage"
     }
 }
