@@ -7,17 +7,18 @@ use Gskema\TypeSniff\Core\Type\TypeInterface;
 
 abstract class AbstractFqcnPropElement extends AbstractFqcnElement
 {
-    /** @var string */
-    protected $propName;
+    protected string $propName;
 
-    /** @var TypeInterface|null */
-    protected $defaultValueType;
+    protected TypeInterface $type;
+
+    protected ?TypeInterface $defaultValueType;
 
     /**
      * @param int                $line
      * @param DocBlock           $docBlock
      * @param string             $fqcn
      * @param string             $propName
+     * @param TypeInterface      $type
      * @param TypeInterface|null $defaultValueType
      * @param string[]           $attributeNames
      */
@@ -26,17 +27,24 @@ abstract class AbstractFqcnPropElement extends AbstractFqcnElement
         DocBlock $docBlock,
         string $fqcn,
         string $propName,
+        TypeInterface $type,
         ?TypeInterface $defaultValueType,
         array $attributeNames
     ) {
         parent::__construct($line, $docBlock, $fqcn, $attributeNames);
         $this->propName = $propName;
+        $this->type = $type;
         $this->defaultValueType = $defaultValueType;
     }
 
     public function getPropName(): string
     {
         return $this->propName;
+    }
+
+    public function getType(): TypeInterface
+    {
+        return $this->type;
     }
 
     public function getDefaultValueType(): ?TypeInterface
