@@ -6,17 +6,13 @@ use Gskema\TypeSniff\Core\Type\TypeInterface;
 
 class VarTag implements TagInterface
 {
-    /** @var int */
-    protected $line;
+    protected int $line;
 
-    /** @var TypeInterface */
-    protected $type;
+    protected TypeInterface $type;
 
-    /** @var string|null */
-    protected $paramName;
+    protected ?string $paramName;
 
-    /** @var string|null */
-    protected $description;
+    protected ?string $description;
 
     public function __construct(
         int $line,
