@@ -51,10 +51,8 @@ class DocTypeInspector
                 $subject->addFnTypeWarning('Missing PHPDoc tag for :subject:');
             }
         } elseif ($subject instanceof PropTypeSubject) {
-            // @var tags for props are mandatory
-            if (!$hasDocTypeTag && !$hasArrayShape) {
-                $subject->addDocTypeWarning('Add @var tag for :subject:');
-            } elseif ($subject->getDocType() instanceof UndefinedType) {
+            // properties: ask to add fn type first
+            if ($subject->getDocType() instanceof UndefinedType) {
                 $subject->addDocTypeWarning('Add type hint to @var tag for :subject:');
             }
         }
@@ -89,15 +87,26 @@ class DocTypeInspector
             return;
         }
 
+        $docType = $subject->getDocType();
+
         // e.g. @param array $arg1 -> @param int[] $arg1
-        if (TypeHelper::containsType($subject->getDocType(), ArrayType::class)) {
+        if (TypeHelper::containsType($docType, NullableType::class)) {
+            $subject->addDocTypeWarning(sprintf(
+                'Replace nullable type "%s" with compound type with null "%s" for :subject:.',
+                $docType->toString(),
+                str_replace('?', '', $docType->toString()) . '|null' // @TODO Not ideal
+            ));
+        }
+
+        // e.g. @param array $arg1 -> @param int[] $arg1
+        if (TypeHelper::containsType($docType, ArrayType::class)) {
             $subject->addDocTypeWarning(
                 'Replace array type with typed array type in PHPDoc for :subject:, .e.g.: "string[]" or "SomeClass[]". Use mixed[] for generic arrays. Correct array depth must be specified.'
             );
         }
 
         // e.g. array[] -> mixed[][]
-        if ($fakeType = TypeHelper::getFakeTypedArrayType($subject->getDocType())) {
+        if ($fakeType = TypeHelper::getFakeTypedArrayType($docType)) {
             $subject->addDocTypeWarning(sprintf(
                 'Use a more specific type in typed array hint "%s" for :subject:. Correct array depth must be specified.',
                 $fakeType->toString()
@@ -164,13 +173,15 @@ class DocTypeInspector
         }
 
         // e.g. ?int, int|string -> ?int, int|null (wrong: string, missing: null)
+        $isProp = $subject instanceof PropTypeSubject;
         [$wrongDocTypes, $missingDocTypes] = TypeComparator::compare(
             $subject->getDocType(),
             $subject->getFnType(),
-            $subject->getValueType()
+            $subject->getValueType(),
+            $isProp
         );
 
-        if ($subject instanceof PropTypeSubject) {
+        if ($isProp && !$subject->hasDefinedFnType()) {
             $wrongDocTypes = []; // not reported because props have dynamic values
         }
 
