@@ -4,7 +4,7 @@ namespace Gskema\TypeSniff\Core\CodeElement\fixtures;
 
 interface TestInterface0
 {
-    const C1 = 1;
+    public const C1 = 1;
 
     public function func1(): void;
 }
